+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Fix displaying default image in gridder
 * Fix gridder not working on pagination
 * Remove googlemaps streetview fallback image from all widgets
+* Disable mousewheel zoom on maps
 
 ## v0.13.0 (2020-02-23)
 * Add login required to choices-guide results